refactor(controllers): extract nested endpoints into separate classes (#86)
